Improved formatting of the footer
