+++ app/(main)/page.tsx
@@ -1,5 +1,5 @@
 import MainContent from "@/app/(main)/_components/main-content";
-import Footer from "@/components/footer";
+import Footer from "@/components/footer/footer";
 import Navbar from "@/components/navbar";
 
 export default function Home() {
